requirements.txt
fastapi
uvicorn[standard]
python-dotenv
firebase-admin
google-cloud-firestore
